requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
faster-whisper
librosa
numpy
requests
langchain
langchain-groq
sqlalchemy

streamlit
pandas
chromadb
sentence-transformers

pyannote.audio
torch
torchvision
torchaudio
soundfile
langchain-text-splitter
schedule
feedparser